feat: add LocationService

+++ src/main/java/com/craftelix/weatherviewer/dto/LocationRequestDto.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -11,8 +13,8 @@ public class LocationRequestDto {
 
     private String name;
 
-    private double latitude;
+    private BigDecimal latitude;
 
-    private double longitude;
+    private BigDecimal longitude;
 
 }
